requirements.txt
streamlit
google-generativeai
langchain
langchain-google-genai
langchain-community
requests